fix: rename + typegoose
