+++ package.json
@@ -19,6 +19,7 @@
     "mock:server": "json-server ./mocks/mock-server-data.json --port 1337"
   },
   "devDependencies": {
+    "@typegoose/typegoose": "11.1.0",
     "@types/convict": "6.1.1",
     "@types/convict-format-with-validator": "6.0.2",
     "@types/node": "18.16.3",
@@ -38,8 +39,6 @@
     "npm": ">=8"
   },
   "dependencies": {
-    "@typegoose/typegoose": "11.1.0",
-    "chalk": "5.2.0",
     "convict": "6.2.4",
     "convict-format-with-validator": "6.2.0",
     "dayjs": "1.11.7",
